Not ready for review yet, I need to write some unit-tests for the `backwards_thru_file` fn (since it doesn't currently have any). Just want to see what CI thinks of it and if it blows up...

These are the performance gains I see...

```
$ hyperfine "/usr/bin/tail -n +100000 ./shakespeare.txt" "./target/release/tail.original -n +100000 ./shakespeare.txt" "./target/release/tail -n +100000 ./shakespeare.txt"
Benchmark 1: /usr/bin/tail -n +100000 ./shakespeare.txt
  Time (mean ± σ):       4.1 ms ±   0.3 ms    [User: 2.1 ms, System: 2.0 ms]
  Range (min … max):     3.8 ms …   5.6 ms    597 runs
 
Benchmark 2: ./target/release/tail.original -n +100000 ./shakespeare.txt
  Time (mean ± σ):       5.4 ms ±   0.3 ms    [User: 3.8 ms, System: 1.5 ms]
  Range (min … max):     4.9 ms …   6.3 ms    461 runs
 
Benchmark 3: ./target/release/tail -n +100000 ./shakespeare.txt
  Time (mean ± σ):       3.2 ms ±   0.3 ms    [User: 1.8 ms, System: 1.4 ms]
  Range (min … max):     2.8 ms …   4.9 ms    704 runs
 
Summary
  ./target/release/tail -n +100000 ./shakespeare.txt ran
    1.31 ± 0.13 times faster than /usr/bin/tail -n +100000 ./shakespeare.txt
    1.69 ± 0.16 times faster than ./target/release/tail.original -n +100000 ./shakespeare.txt
```
and
```
$ hyperfine "/usr/bin/tail -n 100000 ./shakespeare.txt" "./target/release/tail.original -n 100000 ./shakespeare.txt" "./target/release/tail -n 100000 ./shakespeare.txt"
Benchmark 1: /usr/bin/tail -n 100000 ./shakespeare.txt
  Time (mean ± σ):       4.6 ms ±   0.4 ms    [User: 2.4 ms, System: 2.1 ms]
  Range (min … max):     4.2 ms …   7.2 ms    430 runs
 
Benchmark 2: ./target/release/tail.original -n 100000 ./shakespeare.txt
  Time (mean ± σ):       4.4 ms ±   0.3 ms    [User: 2.9 ms, System: 1.5 ms]
  Range (min … max):     3.9 ms …   6.7 ms    595 runs
 
Benchmark 3: ./target/release/tail -n 100000 ./shakespeare.txt
  Time (mean ± σ):       3.4 ms ±   0.3 ms    [User: 2.1 ms, System: 1.3 ms]
  Range (min … max):     3.0 ms …   5.1 ms    715 runs
 
Summary
  ./target/release/tail -n 100000 ./shakespeare.txt ran
    1.28 ± 0.14 times faster than ./target/release/tail.original -n 100000 ./shakespeare.txt
    1.35 ± 0.16 times faster than /usr/bin/tail -n 100000 ./shakespeare.txt
```

So not an order-of-magnitude gain, but a nice little boost none the less.